requirements.txt
absl-py==0.11.0
anytree==2.8.0
astroid==2.4.2
autopep8==1.5.5
backcall==0.2.0
cachelib==0.1.1
certifi==2020.6.20
cffi==1.14.5
click==7.1.2
coclust==0.2.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Script==2.0.6
Flask-Session==0.3.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
k-means-constrained==0.5.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.35.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.5
numba==0.52.0
numpy==1.19.2
opencv-python==4.1.0.25
openpyxl==3.0.5
ortools==8.1.8487
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
pylint==2.6.0
pynndescent==0.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
simplejson==3.17.2
six==1.15.0
some-package==0.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.0
traitlets==5.0.5
typed-ast==1.4.1
ujson==4.0.1
umap-learn==0.5.0
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
lapjv==1.3.1